+++ src/Attributes/WhenLoaded.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace StdOut\SimpleDataObjects\Attributes;
+
+use Attribute;
+
+#[Attribute(Attribute::TARGET_PARAMETER | Attribute::TARGET_PROPERTY)]
+final class WhenLoaded
+{
+    public function __construct(public readonly string $relation) {}
+}